requirements.txt
openai>=1.45.0
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.8.0
PyYAML>=6.0.1
jsonschema>=4.23.0
tqdm>=4.66.0
scikit-learn>=1.5.0
scipy>=1.14.0
networkx>=3.3
matplotlib>=3.9.0
seaborn>=0.13.2
sentence-transformers>=3.0.1
causal-learn>=0.1.3.8
umap-learn>=0.5.6
statsmodels>=0.14.2
transformers>=4.44.0
torch>=2.4.0
